requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
pip==21.1.3
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
wells==1.5.0
wheel==0.36.2
